Remove senior pool v2 page

+++ packages/subgraph/reset-start-local.sh
@@ -1,5 +1,14 @@
 set -e
 
+root_dir=$(dirname $(dirname $(dirname $(readlink -f "$0"))))
+env_path="${root_dir}/.env.local"
+source "${env_path}"
+
+if [ "$REACT_APP_TOGGLE_THE_GRAPH" != "true" ]; then
+    echo "The graph feature flag is toggled off. Exiting..."
+    exit 0
+fi
+
 docker-compose down -v;
 
 echo "Building the sugraph..."
